requirements.txt
pycryptodome>=3.20.0
playwright>=1.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openpyxl>=3.1.0
